Introduce imports.DefaultGoMod
Prior to this change, there was no way for a gomacro host to customise
the generation of the go.mod file by genimport.  This becomes important
when the gomacro host has been compilied with modules fixed at a
particular version and the interpreted gomacro code also makes
refenerence to the same modules.  When genimport compiles it generated
module it will automatically fetch the latest version of the modules in
question, which may not be the version that the host was compiled with.

This change simply introduces some global state that genimport will use
to generate all the go.mod files.  In this way a gomacro host may
specify the modules and their versions that the host itself makes
reference to.